refactor: extract RoutineCoachViewModel and WidgetBridge from ContentViewModel
ContentViewModel pasó de 374 a 273 líneas, sacando dos responsabilidades que no
compartían nada con el resto:

- RoutineCoachViewModel: el coach de IA (Feature 12) con su sugerencia, el estado
  de streaming y el cache por día. Las entries se pasan por parámetro en vez de
  guardarse, así no queda una segunda copia de la lista para mantener en sync.
- WidgetBridge: la definición única del contrato del App Group. Antes el suite
  name y las 3 claves widget* eran literales duplicados en ContentViewModel y en
  MuscleCheckWidget — un typo de un lado rompía el widget en silencio. Ahora los
  dos targets compilan el mismo archivo vía membershipExceptions.

De paso, SharedMuscleEntry estaba duplicado en los dos targets: queda una sola
copia en MuscleCheck/models/, compartida por el mismo mecanismo.
